backend/requirements.txt
selenium
webdriver-manager
beautifulsoup4
flask
anthropic
pymongo
python-dotenv
certifi
flask-cors
